requirements.txt
pytrends
huggingface_hub
jinja2
markdown
pillow
python-slugify
